roi_stats/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-datatables==0.7.1
gunicorn==21.2.0
packaging==24.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
